This PR does not change the functionality of the methods `compress` and `expand`, but improves the underlying computations.

The Base64 <-> Hex conversions here rely on standard python libraries/methods (`base64.b64encode` and `base64.b64decode` as well as `bytes.fromhex`) instead of manual computations.  This makes the algorithms more transparent compared to the previous state. This should also entail a performance boost, though that has yet to be tested.

In addition we have provided docstrings and added references to Ifc documentation.