@Moult 2025-01-19
I'm not entirely sure how this change is a benefit? The previous code seems pretty clear and efficient to me. Am I missing something? Ping @aothms.