@raj-open 2025-01-19
> I'm not entirely sure how this change is a benefit? The previous code seems pretty clear and efficient to me. Am I missing something? Ping @aothms.

Qualitatively there is a significant different between the old an new set of methods. The latter is much more transparent and robust.

Empirically, there is a **significant performance boost**. Using the `codetiming` module, I have ran the old and new methods and the differences are as follows:

| old/new | method | `n` = number of operations | time (seconds) |
| :------ | :----- | -------------------------: | -------------: |
| old | `compress` | 1e6 | 5.0251 |
| new | `compress` | 1e6 | 1.0123 |
| old | `expand` | 1e6 | 7.0914 |
| new | `expand` | 1e6 | 0.5376 |

I would encourage you to test this for yourselves.

(Pseudo)code used for comparison:
```py
from uuid import uuid4
from codetiming import Timer
# import original compress as "old_compress"
# import original expand as "old_expand"
# import new compress as "new_compress"
# import new expand as "new_expand"

@Timer()
def old_compress(uuids: list[str], /):
    for uuid in uuids:
        old_compress(uuid)

@Timer()
def old_expand(uuids: list[str], /):
    for uuid in uuids:
        old_expand(uuid)
@Timer()
def new_compress(uuids: list[str], /):
    for uuid in uuids:
        new_compress(uuid)

@Timer()
def new_expand(uuids: list[str], /):
    for uuid in uuids:
        new_expand(uuid)

# prepare inputs separately
n = int(1e6)
uuids = [str(uuid4().hex) for _ in range(n)]
guids = list(map(compress, uuids))

# run and compare methods
old_compress(uuids)
new_compress(uuids)
old_expand(guids)
new_expand(guids)
```