@aothms 2025-01-20
Couple requests:

- `b64encode` / decode accept an `altchars` kwarg, no reason to use `maketrans`

https://docs.python.org/3/library/base64.html#base64.b64encode

- We use pytest generally instead of `from unittest import TestCase`. You can just remove all the TestCase boilerplate and create functions that start with `test_` for pytest to pick those up.

- Remove unrelated changes from the PR

- The tests are nice, but it would be good to have some formal assurance that the approach for padding is compliant with the spec.