@raj-open 2025-01-20
Hi @aothms! Thank you for looking at the request!

> Couple requests:
> 
> * `b64encode` / decode accept an `altchars` kwarg, no reason to use `maketrans`

I am aware of this. However note that this "altchars" argument is _only_ for the final 2 characters in the base64-alphabet:

> Optional altchars must be a [bytes-like object](https://docs.python.org/3/glossary.html#term-bytes-like-object) **of length 2 which specifies an alternative alphabet for the + and / characters**. This allows an application to e.g. generate URL or filesystem safe Base64 strings. The default is None, for which the standard Base64 alphabet is used.

In the case of IFC's convention, it's not merely these two characters, but the entire alphabet (cf. #5991), which deviates from the global standard of Base 64 encoding. This is the reason why something like `maketrans` is necessary.

> * We use pytest generally instead of `from unittest import TestCase`. You can just remove all the TestCase boilerplate and create functions that start with `test_` for pytest to pick those up.

Oh, my apologies. I will update this. I take it then, that I can completely do away with classes and simple write functions?

> * Remove unrelated changes from the PR

Will do! (I ideally didn't want to add this commit at all : )

> * The tests are nice, but it would be good to have some formal assurance that the approach for padding is compliant with the spec.

Okay, I will think of something. If you have any particular cases/qualities you would like me to check, I can add these.

Briefly: What I originally attempted was to properly pad (on the right with `=` as per standard). However, this led to differing result. I then reflected on IFC's algorithm and realised that it chops off the first 2 characters (in base 64) and converts/pads these. The padding that occurs in `expand` corresponds to this peculiarity. The padding in `compress` simply undoes this effect, so is not special in itself.

> [!NOTE]
> I need some time and will get on to this later this evening (UTC+1).