@raj-open 2025-01-20
> * 6 bytes at a time results in groups of 4 6-byte chars that align with 3 8-byte chars as 4 * 6 == 3 * 8
> * encoded this means 2 | 4 | 4 | 4 | 4 | 4 == 22
> * decoded this means 1 | 3 | 3 | 3 | 3 | 3 == 16 (or double in hex)
> * so that's why you leftpad with 2 bytes '0000' in hex
> * and leftpad with 'AA' in base64 to get to the right number of elements per group so the left to right the marks correspond
> * after leftpadding the corresponding number of chars need to be sliced off the front

Yep, that covers the numerical side of things. The fiddly bit was realising that IFC pads left compared to the usual right-padding under the Base64 convention.

Please let me know if the tests are okay (I ran them separately and they were green; but I could not run them within the IfcOpenShell repository as I cannot compile ifc_wrapper, so I was unable to check if everything is compatible import-wise, etc.).

> I like how you came up with a general solution to something specific 👍

Thanks! It's _almost_ general. The 2 `AA` values / 4 `0000` values thing is a bit of magic (for hex-codes of length `n ≠ 32` this relation can slightly differ), but nonetheless the structure 

```mermaid
flowchart LR
    A["UUID.hex"]
    -->|"left-pad _m_ 0's"| B["36 chars"]
    -->|"encode"| C["UUID.64 (24 chars)"]
    -->|"left-depad _n_ chars"| D["UUID.64 (22 chars)"]
    -->|"translate"| E["UUID.64 ifc"]
```

```mermaid
flowchart LR
    A["UUID.64 ifc"]
    -->|"translate"| B["UUID.64"]
    -->|"left-pad _n_ As"| C["UUID.64 (24 chars)"]
    -->|"decode"| D["UUID.hex (36 chars)"]
    -->|"left-depad _m_ chars"| E["UUID.hex (32 chars)"]
```

appears to capture the essence of IFC's algorithms.