@raj-open 2025-01-25
@aothms it appears that there are issues with the underlying C++ beyond, causing the build step in the ci pipeline to fail. I do not thing it has anything to do with the two python scripts in this PR. Is is possible that something has been merged to v0.8.0 in a broken state?