@aothms 2025-01-28
Yes, you're right. Easy to track down with the runtime of our CI action, which suddenly dropped from 20m to 2m to fail.

![afbeelding](https://github.com/user-attachments/assets/6eedfdf6-00ae-4c02-adb7-e7a91e0bca83)

    git log d1ad082~..f0e1be765d9b2c80e39020b21503b37f7ee2064e --name-only

Doesn't show anything related to swig though so the error is a bit unexplained

    swig4.0: ../../Source/CParse/templ.c:582: Swig_cparse_template_expand: Assertion `!SwigType_isvariadic(s)' failed.

But we can merge this in and resolve the build as part of the other issue you created.

Thanks again!